+++ clang/lib/Sema/SemaDecl.cpp
@@ -10219,6 +10219,7 @@ Sema::ActOnFunctionDeclarator(Scope *S, Declarator &D, DeclContext *DC,
   if (D.isFunctionDefinition()) {
     AddRangeBasedOptnone(NewFD);
     AddImplicitMSFunctionNoBuiltinAttr(NewFD);
+    AddSectionMSAllocText(NewFD);
   }
 
   // If this is the first declaration of an extern C variable, update